\name{R4CDISC-package}
\alias{R4CDISC-package}
\alias{R4CDISC}
\docType{package}
\title{
Read CDISC data files
}
\description{
R4CDISC package includs functions for reading SDS-XML and Define.xml files.
}
\details{
\tabular{ll}{
Package: \tab R4CDISC\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-12-31\cr
License: \tab LGL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Ippei Akiya

Maintainer: Ippei Akiya <yourfault@somewhere.net>
}
\references{

}

\keyword{ CDISC }
\examples{
}
