\name{R4CDISC-package}
\alias{R4CDISC-package}
\alias{R4CDISC}
\docType{package}
\title{
Read CDISC data files
}
\description{
R4CDISC package includs functions for reading SDS-XML and Define.xml files.
}
\details{
\tabular{ll}{
Package: \tab R4CDISC\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-12-31\cr
License: \tab LGL-3\cr
}
R4CDISC package is provide functions to import SDS-XML and define.xml. SDS-XML is able to select  typed or untyped data format, however, R4CDISC package supports only untyped data format.\cr
\cr
It is supported SDS-XML ver. 1.0 draft and define.xml ver. 2.0 in this package.
}
\source{
R4CDISC package includes example files both of SDS-XML and define-xml which were obtained from SDS-XML 1.0 DRAFT package on CDISC web site. \url{http://www.cdisc.org/studydataset-xml}
}
\author{
Ippei Akiya\cr
Maintainer: Ippei Akiya <ippei.aki@datadriven.jp>
}
\references{
CDISC StudyDataSet-XML Specification Version 1.0 Draft (\url{http://www.cdisc.org/studydataset-xml})\cr
CDISC Define-XML Specification Version 2.0 (\url{http://www.cdisc.org/define-xml})
}

\keyword{ CDISC }
